+++ page-landing.php
@@ -38,15 +38,19 @@ get_header();
 
 
 		<section class="content-section entry-content container">
-				<?php
-					the_content( __( 'Continue reading <span class="meta-nav">&rarr;</span>', 'haste-store' ) );
-					wp_link_pages( array(
-						'before'      => '<div class="page-links"><span class="page-links-title">' . __( 'Pages:', 'haste-store' ) . '</span>',
-						'after'       => '</div>',
-						'link_before' => '<span>',
-						'link_after'  => '</span>',
-					) );
-				?>
+			<div class="row">
+				<div class="<?php echo odin_classes_page_full(); ?>">
+					<?php
+						the_content( __( 'Continue reading <span class="meta-nav">&rarr;</span>', 'haste-store' ) );
+						wp_link_pages( array(
+							'before'      => '<div class="page-links"><span class="page-links-title">' . __( 'Pages:', 'haste-store' ) . '</span>',
+							'after'       => '</div>',
+							'link_before' => '<span>',
+							'link_after'  => '</span>',
+						) );
+					?>
+				</div>
+			</div>
 		</section><!-- .entry-content -->
 
 	</article>